refactor: Convert functions to arrow function to follow pattern
refactor: Rename secureStore utility function to follow pattern
