+++ src/utils/secureStoreHelper.ts
@@ -1,18 +1,18 @@
 import { ITokens } from "@type/auth";
 import * as SecureStore from "expo-secure-store";
 
-async function saveToken(key: ITokens, value: string) {
+const saveToken = async (key: ITokens, value: string) => {
   await SecureStore.setItemAsync(key, value);
-}
+};
 
-async function getToken(key: ITokens) {
+const getToken = async (key: ITokens) => {
   let token = await SecureStore.getItemAsync(key);
   return token;
-}
+};
 
-async function deleteToken(key: ITokens) {
+const deleteToken = async (key: ITokens) => {
   await SecureStore.deleteItemAsync(key);
-}
+};
 
 export const secureStore = {
   saveToken,
